Implement step run command execution in executeStep
