+++ cli/step_activator.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/bitrise-io/bitrise/log"
 	"github.com/bitrise-io/bitrise/models"
+	"github.com/bitrise-io/bitrise/tools"
 	"github.com/bitrise-io/go-utils/command"
 	"github.com/bitrise-io/go-utils/command/git"
 	"github.com/bitrise-io/go-utils/pathutil"
@@ -133,4 +134,67 @@ even if the repository is open source!`)
 	}
 
 	return stepYMLPth, origStepYMLPth, nil
+}
+
+func activateStepLibStep(stepIDData models.StepIDData, destination, stepYMLCopyPth string, isStepLibUpdated bool) (stepmanModels.StepInfoModel, bool, error) {
+	didStepLibUpdate := false
+
+	log.Debugf("[BITRISE_CLI] - Steplib (%s) step (id:%s) (version:%s) found, activating step", stepIDData.SteplibSource, stepIDData.IDorURI, stepIDData.Version)
+	if err := tools.StepmanSetup(stepIDData.SteplibSource); err != nil {
+		return stepmanModels.StepInfoModel{}, false, err
+	}
+
+	versionConstraint, err := stepmanModels.ParseRequiredVersion(stepIDData.Version)
+	if err != nil {
+		return stepmanModels.StepInfoModel{}, false,
+			fmt.Errorf("activating step (%s) from source (%s) failed, invalid version specified: %s", stepIDData.IDorURI, stepIDData.SteplibSource, err)
+	}
+	if versionConstraint.VersionLockType == stepmanModels.InvalidVersionConstraint {
+		return stepmanModels.StepInfoModel{}, false,
+			fmt.Errorf("activating step (%s) from source (%s) failed, version constraint is invalid", stepIDData.IDorURI, stepIDData.SteplibSource)
+	}
+
+	isStepLibUpdateNeeded := (versionConstraint.VersionLockType == stepmanModels.Latest) ||
+		(versionConstraint.VersionLockType == stepmanModels.MinorLocked) ||
+		(versionConstraint.VersionLockType == stepmanModels.MajorLocked)
+	if !isStepLibUpdated && isStepLibUpdateNeeded {
+		log.Print("Step uses latest version, updating StepLib...")
+		if err := tools.StepmanUpdate(stepIDData.SteplibSource); err != nil {
+			log.Warnf("Step version constraint is latest or version locked, but failed to update StepLib, err: %s", err)
+		} else {
+			didStepLibUpdate = true
+		}
+	}
+
+	info, err := tools.StepmanStepInfo(stepIDData.SteplibSource, stepIDData.IDorURI, stepIDData.Version)
+	if err != nil {
+		if isStepLibUpdated {
+			return stepmanModels.StepInfoModel{}, didStepLibUpdate, fmt.Errorf("stepman JSON steplib step info failed: %s", err)
+		}
+
+		// May StepLib should be updated
+		log.Infof("Step info not found in StepLib (%s) -- Updating ...", stepIDData.SteplibSource)
+		if err := tools.StepmanUpdate(stepIDData.SteplibSource); err != nil {
+			return stepmanModels.StepInfoModel{}, didStepLibUpdate, err
+		}
+
+		didStepLibUpdate = true
+
+		info, err = tools.StepmanStepInfo(stepIDData.SteplibSource, stepIDData.IDorURI, stepIDData.Version)
+		if err != nil {
+			return stepmanModels.StepInfoModel{}, didStepLibUpdate, fmt.Errorf("stepman JSON steplib step info failed: %s", err)
+		}
+	}
+
+	if info.Step.Title == nil || *info.Step.Title == "" {
+		info.Step.Title = pointers.NewStringPtr(info.ID)
+	}
+	info.OriginalVersion = stepIDData.Version
+
+	if err := tools.StepmanActivate(stepIDData.SteplibSource, stepIDData.IDorURI, info.Version, destination, stepYMLCopyPth); err != nil {
+		return stepmanModels.StepInfoModel{}, didStepLibUpdate, err
+	}
+	log.Debugf("[BITRISE_CLI] - Step activated: (ID:%s) (version:%s)", stepIDData.IDorURI, stepIDData.Version)
+
+	return info, didStepLibUpdate, nil
 }
